+++ examples/_menu.php
@@ -1,43 +1,43 @@
 <ul>
     <li>
-        <a href="bootstrap.php">Bootstrap (5.3.0-alpha1)</a>
+        <a href="bootstrap.php">Bootstrap (5.3.8)</a>
         [<a target="_blank" href="https://getbootstrap.com">https://getbootstrap.com</a>]
     </li>
     <li>
-        <a href="foundation.php">Foundation (6.7.5)</a>
+        <a href="foundation.php">Foundation (6.9.0)</a>
         [<a target="_blank" href="https://get.foundation">https://get.foundation</a>]
     </li>
     <li>
-        <a href="./bulma.php">Bulma (0.9.4)</a>
+        <a href="./bulma.php">Bulma (1.0.4)</a>
         [<a target="_blank" href="https://bulma.io">https://bulma.io</a>]
     </li>
     <li>
-        <a href="./materialize.php">Materialize (1.0.0)</a>
-        [<a target="_blank" href="https://materializecss.com">https://materializecss.com</a>]
-    </li>
-    <li>
-        <a href="./tailwind.php">Tailwind (3.2.4)</a>
+        <a href="./tailwind.php">Tailwind</a>
         [<a target="_blank" href="https://tailwindcss.com">https://tailwindcss.com</a>]
     </li>
     <li>
-        <a href="./primer.php">Primer (20.8.0)</a>
+        <a href="./primer.php">Primer (21.1.1)</a>
         [<a target="_blank" href="https://primer.style">https://primer.style</a>]
     </li>
     <li>
-        <a href="uikit.php">UiKit (3.15.22)</a>
+        <a href="uikit.php">UiKit (3.23.13)</a>
         [<a target="_blank" href="https://getuikit.com">https://getuikit.com</a>]
     </li>
     <li>
-        <a href="semantic.php">Semantic UI (2.5.0)</a>
-        [<a target="_blank" href="https://semantic-ui.com">https://semantic-ui.com</a>]
+        <a href="fomantic.php">Fomantic UI (2.9.4)</a>
+        [<a target="_blank" href="https://fomantic-ui.com">https://fomantic-ui.com</a>]
     </li>
     <li>
-        <a href="./spectre.php">Spectre.css (0.5.9)</a>
-        [<a target="_blank" href="https://picturepan2.github.io/spectre/">https://picturepan2.github.io/spectre/</a>]
+        <a href="./cirrus.php">Cirrus CSS (0.8.0)</a>
+        [<a target="_blank" href="https://cirrus-ui.com">https://cirrus-ui.com</a>]
     </li>
     <li>
-        <a href="./halfmoon.php">Halfmoon (1.1.1)</a>
-        [<a target="_blank" href="https://www.gethalfmoon.com">https://www.gethalfmoon.com</a>]
+        <a href="./vanilla.php">Vanilla Framework (4.21.0)</a>
+        [<a target="_blank" href="https://vanillaframework.io">https://vanillaframework.io</a>]
+    </li>
+    <li>
+        <a href="./beercss.php">Beer CSS (4.0.2)</a>
+        [<a target="_blank" href="https://www.beercss.com">https://www.beercss.com</a>]
     </li>
     <li><a href="./custom.php">Custom Template</a></li>
 </ul>