+++ examples/vanilla.php
@@ -1,0 +1,29 @@
+<?php
+
+use function Tamtamchik\SimpleFlash\flash;
+
+include_once '_init.php';
+?>
+
+<!doctype html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Test Vanilla Framework template example.</title>
+    <!-- Latest compiled and minified CSS -->
+    <link rel="stylesheet" href="https://assets.ubuntu.com/v1/vanilla-framework-version-4.21.0.min.css"/>
+</head>
+<body>
+
+<?php include_once '_ribbon.php'; ?>
+
+<div class="row" style="width: 600px; margin: 0 auto;">
+
+    <?php include_once '_menu.php'; ?>
+
+    <?= flash()->displayVanilla() ?>
+
+</div>
+
+</body>
+</html>