+++ src/Templates.php
@@ -12,11 +12,11 @@ class Templates
     public const BOOTSTRAP = 'bootstrap';
     public const FOUNDATION = 'foundation';
     public const BULMA = 'bulma';
-    public const MATERIALIZE = 'materialize';
     public const TAILWIND = 'tailwind';
     public const PRIMER = 'primer';
     public const UIKIT = 'uikit';
-    public const SEMANTIC = 'semantic';
-    public const SPECTRE = 'spectre';
-    public const HALFMOON = 'halfmoon';
+    public const FOMANTIC = 'fomantic';
+    public const CIRRUS = 'cirrus';
+    public const VANILLA = 'vanilla';
+    public const BEERCSS = 'beercss';
 }