+++ src/Templates/CirrusTemplate.php
@@ -6,18 +6,16 @@ use Tamtamchik\SimpleFlash\BaseTemplate;
 use Tamtamchik\SimpleFlash\TemplateInterface;
 
 /**
- * Class HalfmoonTemplate.
- * Uses default Halfmoon markdown for flash messages.
+ * Class CirrusTemplate.
+ * Uses default Cirrus CSS markup for flash messages.
  */
-class HalfmoonTemplate extends BaseTemplate implements TemplateInterface
+class CirrusTemplate extends BaseTemplate implements TemplateInterface
 {
     protected $prefix = '';
     protected $postfix = '<br />';
-    protected $wrapper = '<div class="alert alert-%s" role="alert">%s</div>';
+    protected $wrapper = '<div class="toast toast--%s">%s</div>';
 
     /**
-     * Override base function to suite Bootstrap 3 alert naming.
-     *
      * @param string $messages - message text
      * @param string $type - message type: success, info, warning, error
      *
@@ -26,8 +24,6 @@ class HalfmoonTemplate extends BaseTemplate implements TemplateInterface
     public function wrapMessages(string $messages, string $type): string
     {
         $type = ($type == 'error') ? 'danger' : $type;
-        $type = ($type == 'warning') ? 'secondary' : $type;
-        $type = ($type == 'info') ? 'primary' : $type;
 
         return sprintf($this->getWrapper(), $type, $messages);
     }